refactor: Replace bcrypt with bcryptjs for better compatibility
- Update package.json to use bcryptjs instead of bcrypt
- Update passwordHasher adapter to import from bcryptjs
- Rename adapter directory from bcrypt to bcryptjs
- Update context.ts import path
